Refactor: Remove old model files and seed data; introduce new models with updated structure
- Deleted outdated Order, Products, Review, Seller, User models and seed data files.
- Added new Cart model to manage user cart items.
- Introduced new Order, Products, Review, Seller, and User models with updated schemas.
- Updated tsconfig to include new model files.
- Refactored seed data to align with new model structure and added sample data for users, sellers, products, cart, orders, and reviews.
