+++ tsconfig.json
@@ -27,6 +27,6 @@
       "@/utils/*": ["utils/*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "src/models/seed-data.js", "src/models/Seller.js", "src/models/Cart.js", "src/models/User.js", "src/models/Review.js", "src/models/Order.js", "src/models/Products.js"],
   "exclude": ["node_modules"]
 }
